requirements.txt
fastapi
uvicorn
garminconnect
pandas
scikit-learn
python-dotenv
sqlalchemy
pydantic

